Unit testing CalculateMealNutrients use case

+++ core/src/main/java/com/aminovic/core/data/preferences/DefaultPreferences.kt
@@ -51,19 +51,19 @@ class DefaultPreferences(
 
     override fun saveCarbRatio(ratio: Int) {
         sharedPref.edit()
-            .putInt(Preferences.KEY_CARB_RATIO, ratio)
+            .putFloat(Preferences.KEY_CARB_RATIO, ratio.toFloat())
             .apply()
     }
 
     override fun saveProteinRatio(ratio: Int) {
         sharedPref.edit()
-            .putInt(Preferences.KEY_PROTEIN_RATIO, ratio)
+            .putFloat(Preferences.KEY_PROTEIN_RATIO, ratio.toFloat())
             .apply()
     }
 
     override fun saveFatRatio(ratio: Int) {
         sharedPref.edit()
-            .putInt(Preferences.KEY_FAT_RATIO, ratio)
+            .putFloat(Preferences.KEY_FAT_RATIO, ratio.toFloat())
             .apply()
     }
 
@@ -74,9 +74,9 @@ class DefaultPreferences(
         val genderString = sharedPref.getString(Preferences.KEY_GENDER, null)
         val activityLevelString = sharedPref.getString(Preferences.KEY_ACTIVITY_LEVEL, null)
         val goalTypeString = sharedPref.getString(Preferences.KEY_GOAL_TYPE, null)
-        val carbRatio = sharedPref.getInt(Preferences.KEY_CARB_RATIO, -1)
-        val proteinRatio = sharedPref.getInt(Preferences.KEY_PROTEIN_RATIO, -1)
-        val fatRatio = sharedPref.getInt(Preferences.KEY_FAT_RATIO, -1)
+        val carbRatio = sharedPref.getFloat(Preferences.KEY_CARB_RATIO, -1f)
+        val proteinRatio = sharedPref.getFloat(Preferences.KEY_PROTEIN_RATIO, -1f)
+        val fatRatio = sharedPref.getFloat(Preferences.KEY_FAT_RATIO, -1f)
 
         return UserInfo(
             gender = Gender.fromString(genderString ?: "male"),
